% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{births}
\alias{births}
\title{Number of births in Sweden 1969-2014}
\format{A list with two matrices, each with 46 rows (ages 14-49) and
46 columns (years 1969-2014). Age 14 is 14- and age 49 is 49+.}
\usage{
data(births)
}
\description{
A list containing two matrices, 'males' and 'females'. Each matrix shows the
 number of births by mother's age and year
}
\keyword{datasets}

