% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deaths.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{Number of deaths in Sweden 1969-2014}
\format{A list with two matrices, each with 101 rows (ages 0-100) and
46 columns (years 1969-2014). Age 100 is in fact 100 and above.}
\description{
A list containing two matrices, 'males' and 'females'. Each matrix shows the
 number of deaths by age and year
}
\keyword{datasets}

