% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBirths.R
\name{getBirths}
\alias{getBirths}
\title{Extract the number of births in Sweden by age, year, and sex.}
\usage{
getBirths(sex = c("all", "males", "females"), years = 1969:2014,
  ages = 14:49, aggrYear = FALSE, aggrAge = FALSE)
}
\arguments{
\item{sex}{"all", "males", or "females"}

\item{years}{years to include, numeric vector.}

\item{ages}{ages to include, numeric vector.}

\item{aggrYear}{logical, if TRUE aggregate ove years.}

\item{aggrAge}{logical, if TRUE aggregate over ages.}
}
\value{
A matrix, a vector, or a real number, depending on aggregate level.
}
\description{
Extract the number of births in Sweden by age, year, and sex.
}
\examples{
getBirths()
getBirths(sex = "males", years = 1969:1978, aggrYear = TRUE)
}

