% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDeaths.R
\name{getDeaths}
\alias{getDeaths}
\title{Extract the number of deaths in Sweden by age, year, and sex.}
\usage{
getDeaths(sex = c("all", "males", "females"), years = 1969:2014,
  ages = 0:100, aggrYear = FALSE, aggrAge = FALSE)
}
\arguments{
\item{sex}{"all", "males", or "females"}

\item{years}{years to include, numeric vector.}

\item{ages}{ages to include, numeric vector.}

\item{aggrYear}{logical, if TRUE aggregate ove years.}

\item{aggrAge}{logical, if TRUE aggregate over ages.}
}
\value{
A matrix, a vector, or a real number, depending on aggregate level.
}
\description{
Extract the number of deaths in Sweden by age, year, and sex.
}
\examples{
getDeaths()
getDeaths(sex = "males", years = 1969:1978, aggrYear = TRUE)
}

