% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perc.R
\name{perc}
\alias{perc}
\title{Calculates a quantile.}
\usage{
perc(value, pop, ages = 0:100)
}
\arguments{
\item{value}{A brobability in percent}

\item{pop}{a matrix of population by age and year}

\item{ages}{vector of ages to consider}
}
\value{
A (vector of) quantile(s)
}
\description{
Calculates a quantile.
}
\examples{
women <- getPop(sex = "females")
perc(50, women)
}

