% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Calculates the percentile of an age distribution}
\usage{
prop(age, pop, below = TRUE, ages = 0:100)
}
\arguments{
\item{age}{The quantile.}

\item{pop}{A matrix of population data.}

\item{below}{logical, if true the lower tail is calculated.}

\item{ages}{A vector of ages to consider.}
}
\value{
A (vector of) percentile(s).
}
\description{
Calculates the percentile of an age distribution
}
\examples{
men <- getPop(sex = "males")
prop(65, men, below = FALSE)
}

